requirements.txt
fastapi~=0.121.1
SQLAlchemy~=2.0.44
pydantic~=2.12.4
aiohttp~=3.13.2
python-dotenv~=1.2.1